refactor(page): extract useHollowingManager hook (phase 3)
Move the hollowing subsystem — shell/open-face/draft state, voxel-edit
mode + undo/redo, the preview cache/debounce machinery, ~190-line apply,
runHollowPreview, and the hollowing effects — out of page.tsx into
features/hollowing/useHollowingManager.ts (reusing the Phase-1
hollowingSerialize/hollowingPreviewTypes/hollowingPreviewCache modules).

Hollowing and hole-punch are bidirectionally coupled, so the hook is
called early (its state is consumed by the existing useHolePunchManager
call via destructure-same-names) and its late/cross-domain deps —
including values returned by useHolePunchManager — are supplied through a
MutableRefObject<HollowingManagerDeps> populated after both managers
exist. Shared modifier-reset/blocker machinery + view/shader state stay
in Home. 66 names returned. Authored by a worktree subagent, integrated
and re-verified here.

page.tsx: 19,133 -> 17,869 lines (-1,264). Verified via `tsc --noEmit`.
Smoke-test hollowing apply/preview/voxel-edit before merge.
