+++ src/app/page.tsx
@@ -15,6 +15,8 @@ import { PrintingPreviewPane } from '@/components/organisms/PrintingPreviewPane'
 import { useMirrorManager } from '@/features/mirror/useMirrorManager';
 import { useArrangeManager } from '@/features/scene/arrange/useArrangeManager';
 import { useHolePunchManager } from '@/features/hole-punching/useHolePunchManager';
+import { useHollowingManager } from '@/features/hollowing/useHollowingManager';
+import type { HollowingManagerDeps } from '@/features/hollowing/useHollowingManager';
 import { GlobalUpdateIndicator } from '@/features/updater/GlobalUpdateIndicator';
 import { EmptySceneState } from '@/components/layout/EmptySceneState';
 import { IslandScanCard } from '@/components/controls/IslandScanCard';
@@ -600,6 +602,101 @@ export default function Home() {
   // 2. Transform Management (needs geom for bounds)
   const transformMgr = useTransformManager({ geom: scene.geom });
 
+  // --- Hollowing manager: placed early so its state/setters are in scope for
+  //     useHolePunchManager below. Late/cross deps supplied via a ref populated
+  //     after the hole-punch manager and shared callbacks exist (TDZ break). ---
+  const hollowingDepsRef = React.useRef<HollowingManagerDeps>({
+    showOperationError: () => {},
+    setShowDamagedModelDialog: () => {},
+    persistActiveModelModifiers: () => {},
+    setPendingModifierResetAction: () => {},
+    setInteriorView: () => {},
+    setSessionShaderOverride: () => {},
+    computeAutoHolePunchDepthMmForGeometry: () => 0,
+    setHolePunchState: () => {},
+    setHolePunchPlacements: () => {},
+    holePunchPlacementsRef: { current: [] },
+    setPendingHolePunchAutoApplyModelId: () => {},
+    setPendingBlockerResetState: () => {},
+    setSelectedHolePunchPlacementIds: () => {},
+    setHoveredHolePunchPlacementId: () => {},
+    setHolePunchHoverPlacement: () => {},
+    interiorView: false,
+  });
+  const hollowing = useHollowingManager({
+    scene,
+    transformMgr,
+    deps: hollowingDepsRef,
+  });
+  const {
+    isPreviewingHollowing,
+    setIsPreviewingHollowing,
+    hollowPreview,
+    setHollowPreview,
+    hollowingState,
+    setHollowingState,
+    isShellOpenFaceSelected,
+    setIsShellOpenFaceSelected,
+    hollowingDraftEnabled,
+    setHollowingDraftEnabled,
+    hollowingEditMode,
+    setHollowingEditMode,
+    blockedHollowVoxelIndices,
+    setBlockedHollowVoxelIndices,
+    editingBlockedHollowVoxelIndices,
+    setEditingBlockedHollowVoxelIndices,
+    isApplyingBlockersHollowing,
+    setIsApplyingBlockersHollowing,
+    isApplyingHollowing,
+    setIsApplyingHollowing,
+    editingBlockedHollowVoxelIndicesRef,
+    hollowVoxelEditUndoStackRef,
+    hollowVoxelEditRedoStackRef,
+    hollowingEditModeRef,
+    hollowPreviewDebounceTimerRef,
+    hollowPreviewRequestSeqRef,
+    hollowPreviewResultCacheRef,
+    hollowPreviewWarmupKeyRef,
+    hollowingSourceByModelIdRef,
+    cavityGeometryByModelIdRef,
+    defaultHollowingState,
+    isHollowingApplied,
+    persistedHollowingSignature,
+    draftHollowingSignature,
+    isShellFaceSelectionPending,
+    isHollowingDirty,
+    canResetHollowing,
+    blockedHollowVoxelIndexSet,
+    editingBlockedHollowVoxelIndexSet,
+    blockedPreviewVoxelInstanceIdSet,
+    handleApplyHollowing,
+    handleResetHollowing,
+    handleClearAppliedHollowing,
+    handleResetHollowingSettings,
+    handleHollowingStateChange,
+    applyEditingBlockedHollowVoxelIndices,
+    undoHollowVoxelEdit,
+    redoHollowVoxelEdit,
+    commitBlockedHollowVoxelIndices,
+    toggleBlockedHollowVoxelIndex,
+    requestResetHollowing,
+    requestClearAppliedHollowing,
+    clearPendingHollowPreviewDebounce,
+    resolveHollowPreviewSourceGeometry,
+    buildHollowingOptions,
+    buildHollowPreviewRequest,
+    cacheHollowPreviewResult,
+    materializeHollowPreviewCacheEntry,
+    primeHollowPreviewCache,
+    runHollowPreview,
+    clearHollowPreview,
+    resolveBlockedHollowVoxelMarqueeSelection,
+    handleBlockedHollowVoxelMarqueeSelection,
+    handleStartHollowVoxelEditing,
+    handleClearHollowVoxelEditing,
+    handleDoneHollowVoxelEditing,
+  } = hollowing;
+
   // Ref for supports group (used for export)
   const supportsRef = React.useRef<THREE.Group | null>(null);
   // Hide support geometry in hollowing mode — it just gets in the way.
@@ -870,50 +967,18 @@ export default function Home() {
 
   const [sessionShaderOverride, setSessionShaderOverride] = React.useState<MeshShaderType | null>(null);
   const [interiorView, setInteriorView] = React.useState(false);
-  const [isPreviewingHollowing, setIsPreviewingHollowing] = React.useState(false);
-  const [hollowPreview, setHollowPreview] = React.useState<HollowPreviewState | null>(null);
-  const shouldForceHollowingXray = scene.mode === 'prepare'
-    && transformMgr.transformMode === 'hollowing'
-    && !scene.activeModel?.meshModifiers?.hollowing?.bakedIntoGeometry;
-  const effectiveShaderType = (shouldForceHollowingXray || hollowPreview)
-    ? 'xray'
-    : (sessionShaderOverride ?? scene.shaderType);
   const [isPrepareDragActive, setIsPrepareDragActive] = React.useState(false);
   const [isPrepareDragUnsupported, setIsPrepareDragUnsupported] = React.useState(false);
   const [isSupportSpotlightHoldActive, setIsSupportSpotlightHoldActive] = React.useState(false);
   const [allowPrepareWithoutPrinter, setAllowPrepareWithoutPrinter] = React.useState(false);
   const [prepareSmoothingSettingsExpanded, setPrepareSmoothingSettingsExpanded] = React.useState(true);
-  const [hollowingState, setHollowingState] = React.useState<HollowingPanelState>({
-    mode: 'cavity',
-    voxelSizeMm: 0.65,
-    shellThicknessMm: 2.0,
-    infillMode: 'lattice',
-    infillCellMm: 4.2426,
-    infillBeamRadiusMm: 0.35,
-    openFace: 'z_max',
-  });
-  const [isShellOpenFaceSelected, setIsShellOpenFaceSelected] = React.useState(true);
-  const [hollowingDraftEnabled, setHollowingDraftEnabled] = React.useState(false);
-  const [hollowingEditMode, setHollowingEditMode] = React.useState(false);
-  const [blockedHollowVoxelIndices, setBlockedHollowVoxelIndices] = React.useState<number[]>([]);
-  const [editingBlockedHollowVoxelIndices, setEditingBlockedHollowVoxelIndices] = React.useState<number[]>([]);
-  const editingBlockedHollowVoxelIndicesRef = React.useRef<number[]>([]);
-  const hollowVoxelEditUndoStackRef = React.useRef<number[][]>([]);
-  const hollowVoxelEditRedoStackRef = React.useRef<number[][]>([]);
-  const hollowingEditModeRef = React.useRef(false);
   const [selectedHolePunchPlacementIds, setSelectedHolePunchPlacementIds] = React.useState<string[]>([]);
   const [hoveredHolePunchPlacementId, setHoveredHolePunchPlacementId] = React.useState<string | null>(null);
   const [holePunchHoverPlacement, setHolePunchHoverPlacement] = React.useState<HolePunchPlacementState | null>(null);
   const [isApplyingHolePunch, setIsApplyingHolePunch] = React.useState(false);
-  const [isApplyingBlockersHollowing, setIsApplyingBlockersHollowing] = React.useState(false);
   const [pendingHolePunchAutoApplyModelId, setPendingHolePunchAutoApplyModelId] = React.useState<string | null>(null);
-  const [isApplyingHollowing, setIsApplyingHollowing] = React.useState(false);
   const [pendingModifierResetAction, setPendingModifierResetAction] = React.useState<PendingModifierResetAction | null>(null);
   const [pendingBlockerResetState, setPendingBlockerResetState] = React.useState<HollowingPanelState | null>(null);
-  const hollowPreviewDebounceTimerRef = React.useRef<number | ReturnType<typeof setTimeout> | null>(null);
-  const hollowPreviewRequestSeqRef = React.useRef(0);
-  const hollowPreviewResultCacheRef = React.useRef<Map<string, HollowPreviewCacheEntry>>(new Map());
-  const hollowPreviewWarmupKeyRef = React.useRef<string | null>(null);
   const [debugPrimitivesPanelVisible, setDebugPrimitivesPanelVisible] = React.useState<boolean>(false);
   const [editorContextMenuPos, setEditorContextMenuPos] = React.useState<{ x: number; y: number } | null>(null);
   const [editorContextMenuSupportTarget, setEditorContextMenuSupportTarget] = React.useState<{
@@ -955,15 +1020,6 @@ export default function Home() {
   const [isPrintingPreviewSettled, setIsPrintingPreviewSettled] = React.useState(false);
   const [isPrintingSettledCanvasReady, setIsPrintingSettledCanvasReady] = React.useState(false);
 
-  const defaultHollowingState = React.useMemo<HollowingPanelState>(() => ({
-    mode: 'cavity',
-    voxelSizeMm: 0.65,
-    shellThicknessMm: 2.0,
-    infillMode: 'lattice',
-    infillCellMm: 4.2426,
-    infillBeamRadiusMm: 0.35,
-    openFace: 'z_max',
-  }), []);
 
   const defaultHolePunchState = React.useMemo<HolePunchPanelState>(() => ({
     radiusMm: 2.0,
@@ -975,8 +1031,6 @@ export default function Home() {
     () => getDefaultHolePunchDepthMm(hollowingState.shellThicknessMm),
     [hollowingState.shellThicknessMm],
   );
-  const hollowingSourceByModelIdRef = React.useRef<Map<string, HollowingSourceEntry>>(new Map());
-  const cavityGeometryByModelIdRef = React.useRef<Map<string, CavityGeometryEntry>>(new Map());
   const [printingPreviewZoom, setPrintingPreviewZoom] = React.useState(1);
   const [printingPreviewPan, setPrintingPreviewPan] = React.useState({ x: 0, y: 0 });
   const [isPrintingPreviewPanning, setIsPrintingPreviewPanning] = React.useState(false);
@@ -13003,701 +13057,35 @@ export default function Home() {
     handleApplyHolePunch,
   } = holePunch;
 
-  const handleApplyHollowing = React.useCallback(() => {
-    void (async () => {
-      const activeModel = scene.activeModel;
-      if (!activeModel) return;
-
-      const persistedHollowing = activeModel.meshModifiers?.hollowing;
-      const shouldApply = hollowingDraftEnabled || !persistedHollowing?.enabled;
-      if (!shouldApply) return;
-
-      if (hollowingState.mode === 'shell_open_face' && !isShellOpenFaceSelected) {
-        showOperationError('Pick the face to open before applying Shell mode.');
-        return;
-      }
-
-      const holesWereAlreadyBaked = activeModel.meshModifiers?.holePunchesBakedIntoGeometry === true;
-
-      const existingSource = hollowingSourceByModelIdRef.current.get(activeModel.id);
-      const needsNewSource = !existingSource || !persistedHollowing?.enabled;
-
-      let sourceGeometry: THREE.BufferGeometry;
-      if (needsNewSource) {
-        if (existingSource) {
-          existingSource.geometry.dispose();
-        }
-
-        if (persistedHollowing?.enabled) {
-          const restoredFromSnapshot = geometryFromSnapshot(persistedHollowing);
-          sourceGeometry = restoredFromSnapshot ?? activeModel.geometry.geometry.clone();
-        } else {
-          // Use the current geometry which may already have baked holes.
-          sourceGeometry = activeModel.geometry.geometry.clone();
-        }
-
-        hollowingSourceByModelIdRef.current.set(activeModel.id, {
-          geometry: sourceGeometry,
-        });
-      } else {
-        sourceGeometry = existingSource.geometry;
-      }
-
-      setIsApplyingHollowing(true);
-      try {
-        const effectiveHollowMode = hollowingState.mode === 'shell_open_face'
-          ? 'cavity'
-          : hollowingState.mode;
-        const shellScaleFactor = getUniformScaleFactorForThickness(activeModel.transform.scale);
-        const bbox = sourceGeometry.boundingBox ?? new THREE.Box3().setFromBufferAttribute(
-          sourceGeometry.getAttribute('position') as THREE.BufferAttribute,
-        );
-        const bboxSize = bbox.getSize(new THREE.Vector3());
-        const maxExtent = Math.max(bboxSize.x, bboxSize.y, bboxSize.z);
-        const applyQuat = new THREE.Quaternion().setFromEuler(activeModel.transform.rotation);
-        const options: HollowOptions = {
-          mode: effectiveHollowMode,
-          voxelResolution: computeVoxelResolution(worldMmToLocalMm(hollowingState.voxelSizeMm, shellScaleFactor), maxExtent),
-          shellThicknessMm: worldMmToLocalMm(hollowingState.shellThicknessMm, shellScaleFactor),
-          blockedVoxelIndices: blockedHollowVoxelIndices,
-          infillMode: hollowingState.infillMode,
-          infillCellMm: worldMmToLocalMm(hollowingState.infillCellMm, shellScaleFactor),
-          infillBeamRadiusMm: worldMmToLocalMm(hollowingState.infillBeamRadiusMm, shellScaleFactor),
-          openFace: hollowingState.openFace,
-          drainHoles: [],
-          previewCavityOnly: false,
-          smoothInternalSurfaces: true,
-          internalChamferPasses: 2,
-          rotationQuat: [applyQuat.x, applyQuat.y, applyQuat.z, applyQuat.w],
-        };
-        const sourceGeometryKey = buildGeometryVersionKey(sourceGeometry);
-        const staged = await stageHollowPreviewSource(
-          sourceGeometry,
-          `${activeModel.id}::${sourceGeometryKey}`,
-        );
-
-        const result = staged
-          ? await hollowApplyFromCapturedSource(options)
-          : await hollowFromGeometry(sourceGeometry, options);
-        if (!result) {
-          showOperationError('Hollowing is available in DragonFruit Desktop only.');
-          return;
-        }
-
-        // Detect hollowing failure: if no voxels were removed, the manifold
-        // stabilization could not resolve the cavity surface, likely because
-        // the mesh is too damaged for boolean operations.
-        if (result.report && 'removedVoxels' in result.report && result.report.removedVoxels === 0) {
-          setShowDamagedModelDialog(true);
-          return;
-        }
-
-        const nextGeometry = new THREE.BufferGeometry();
-        nextGeometry.setAttribute('position', new THREE.BufferAttribute(result.positions, 3));
-        nextGeometry.computeVertexNormals();
-        nextGeometry.computeBoundingBox();
-        nextGeometry.computeBoundingSphere();
-
-        // Store cavity geometry for Interior View Mode
-        if (result.cavityPositions) {
-          const existingCavity = cavityGeometryByModelIdRef.current.get(activeModel.id);
-          if (existingCavity) {
-            existingCavity.geometry.dispose();
-          }
-          const cavityGeometry = new THREE.BufferGeometry();
-          cavityGeometry.setAttribute('position', new THREE.BufferAttribute(result.cavityPositions, 3));
-          cavityGeometry.computeVertexNormals();
-          cavityGeometry.computeBoundingBox();
-          cavityGeometry.computeBoundingSphere();
-          cavityGeometryByModelIdRef.current.set(activeModel.id, { geometry: cavityGeometry });
-        } else {
-          const existingCavity = cavityGeometryByModelIdRef.current.get(activeModel.id);
-          if (existingCavity) {
-            existingCavity.geometry.dispose();
-            cavityGeometryByModelIdRef.current.delete(activeModel.id);
-          }
-        }
-
-        const modeLabel = hollowingState.mode === 'shell_open_face'
-          ? 'Shell Hollowing'
-          : hollowingState.mode === 'infill'
-            ? 'Infill Hollowing'
-            : 'Cavity Hollowing';
-        const replaced = scene.replaceModelGeometry(
-          activeModel.id,
-          nextGeometry,
-          `${modeLabel} (${result.report.outputTriangleCount.toLocaleString()} tris)`,
-        );
-        if (!replaced) {
-          nextGeometry.dispose();
-          return;
-        }
-
-        // Hollowing is now baked — clear the preview overlay and exit X-Ray
-        // forced shader so the user can see surface detail for hole placement.
-        clearHollowPreview();
-        setSessionShaderOverride(null);
-
-        const sourceSnapshot = snapshotGeometryPositions(sourceGeometry);
-        let cavityPositionsBase64: string | undefined;
-        let cavityPositionCount: number | undefined;
-        if (result.cavityPositions) {
-          const cavityBytes = new Uint8Array(
-            result.cavityPositions.buffer,
-            result.cavityPositions.byteOffset,
-            result.cavityPositions.byteLength,
-          );
-          cavityPositionsBase64 = bytesToBase64(cavityBytes);
-          cavityPositionCount = result.cavityPositions.length / 3;
-        }
-
-        setHolePunchState((previous) => (
-          previous.depthMode === 'auto'
-            ? previous
-            : { ...previous, depthMode: 'auto' }
-        ));
-
-        const nextHolePunchPlacements = holePunchPlacementsRef.current.map((placement) => {
-          if (placement.modelId !== activeModel.id || placement.depthMode !== 'auto') {
-            return placement;
-          }
-
-          return {
-            ...placement,
-            depthMm: computeAutoHolePunchDepthMmForGeometry(
-              activeModel,
-              nextGeometry,
-              placement.worldPoint,
-              placement.worldNormal,
-            ),
-          };
-        });
-        setHolePunchPlacements(nextHolePunchPlacements);
-
-        const persistedHolePunches = toPersistedHolePunchPlacements(
-          { geometry: { geometry: nextGeometry } as GeometryWithBounds },
-          nextHolePunchPlacements.filter((placement) => placement.modelId === activeModel.id),
-        ).filter((placement) => placement.radiusMm > 0 && placement.depthMm > 0);
-
-        // When holes were already baked before hollowing, they were passed as
-        // drainHoles to the hollower which already cut them — no re-apply needed.
-        // Only auto-reapply holes that were in draft state (not yet baked).
-        const shouldAutoReapplyHolePunches = !holesWereAlreadyBaked && persistedHolePunches.length > 0;
-
-        persistActiveModelModifiers({
-          ...(activeModel.meshModifiers ?? {}),
-          hollowing: {
-            enabled: true,
-            bakedIntoGeometry: true,
-            sourcePositionsBase64: sourceSnapshot.sourcePositionsBase64,
-            sourcePositionCount: sourceSnapshot.sourcePositionCount,
-            cavityPositionsBase64,
-            cavityPositionCount,
-            blockedVoxelIndices: blockedHollowVoxelIndices,
-            mode: effectiveHollowMode,
-            voxelSizeMm: hollowingState.voxelSizeMm,
-            shellThicknessMm: hollowingState.shellThicknessMm,
-            infillMode: hollowingState.infillMode,
-            infillCellMm: hollowingState.infillCellMm,
-            infillBeamRadiusMm: hollowingState.infillBeamRadiusMm,
-            openFace: hollowingState.openFace,
-            openFaceSelected: hollowingState.mode === 'shell_open_face'
-              ? isShellOpenFaceSelected
-              : true,
-          },
-          holePunches: persistedHolePunches,
-          holePunchAppliedPlacements: holesWereAlreadyBaked ? persistedHolePunches : [],
-          holePunchesBakedIntoGeometry: holesWereAlreadyBaked,
-          holePunchSourcePositionsBase64: holesWereAlreadyBaked
-            ? (activeModel.meshModifiers?.holePunchSourcePositionsBase64 ?? undefined)
-            : undefined,
-          holePunchSourcePositionCount: holesWereAlreadyBaked
-            ? (activeModel.meshModifiers?.holePunchSourcePositionCount ?? undefined)
-            : undefined,
-        });
-
-        if (shouldAutoReapplyHolePunches) {
-          setPendingHolePunchAutoApplyModelId(activeModel.id);
-        }
-      } catch (error) {
-        const message = error instanceof Error ? error.message : String(error);
-        showOperationError(`Hollowing failed: ${message}`);
-      } finally {
-        setIsApplyingHollowing(false);
-        setIsApplyingBlockersHollowing(false);
-      }
-    })();
-  }, [blockedHollowVoxelIndices, hollowingDraftEnabled, hollowingState, isShellOpenFaceSelected, persistActiveModelModifiers, scene]);
-
-  const handleResetHollowing = React.useCallback(() => {
-    const activeModel = scene.activeModel;
-    if (!activeModel) return;
-
-    const sourceEntry = hollowingSourceByModelIdRef.current.get(activeModel.id)
-      ?? (() => {
-        const restored = geometryFromSnapshot(activeModel.meshModifiers?.hollowing ?? {});
-        if (!restored) return null;
-        const entry = { geometry: restored };
-        hollowingSourceByModelIdRef.current.set(activeModel.id, entry);
-        return entry;
-      })();
-
-    if (sourceEntry) {
-      const restoredGeometry = sourceEntry.geometry.clone();
-      const restored = scene.replaceModelGeometry(activeModel.id, restoredGeometry, 'Reset Hollowing');
-      if (!restored) {
-        restoredGeometry.dispose();
-      }
-    }
-
-    // Clear cavity geometry and auto-disable interior view on hollowing reset
-    const existingCavity = cavityGeometryByModelIdRef.current.get(activeModel.id);
-    if (existingCavity) {
-      existingCavity.geometry.dispose();
-      cavityGeometryByModelIdRef.current.delete(activeModel.id);
-    }
-    setInteriorView(false);
-
-    setHollowingState(defaultHollowingState);
-    setIsShellOpenFaceSelected(true);
-    setHollowingDraftEnabled(false);
-    setHollowingEditMode(false);
-    setBlockedHollowVoxelIndices([]);
-    setEditingBlockedHollowVoxelIndices([]);
-    persistActiveModelModifiers({
-      ...(activeModel.meshModifiers ?? {}),
-      hollowing: {
-        enabled: false,
-        bakedIntoGeometry: false,
-        // Clear the source snapshot — hollowing was reset so the snapshot is
-        // stale (it may contain holes that have since been removed).
-        sourcePositionsBase64: undefined,
-        sourcePositionCount: undefined,
-        blockedVoxelIndices: [],
-        mode: defaultHollowingState.mode,
-        voxelSizeMm: defaultHollowingState.voxelSizeMm,
-        shellThicknessMm: defaultHollowingState.shellThicknessMm,
-        infillMode: defaultHollowingState.infillMode,
-        infillCellMm: defaultHollowingState.infillCellMm,
-        infillBeamRadiusMm: defaultHollowingState.infillBeamRadiusMm,
-        openFace: defaultHollowingState.openFace,
-        openFaceSelected: true,
-      },
-      // Preserve hole punch baked state — the geometry restored from the
-      // hollowing source still contains any pre-baked holes, so the system
-      // must not lose track of them.
-      holePunchAppliedPlacements: activeModel.meshModifiers?.holePunches ?? [],
-      holePunchesBakedIntoGeometry: activeModel.meshModifiers?.holePunchesBakedIntoGeometry === true,
-      holePunchSourcePositionsBase64: activeModel.meshModifiers?.holePunchSourcePositionsBase64,
-      holePunchSourcePositionCount: activeModel.meshModifiers?.holePunchSourcePositionCount,
-    });
-  }, [defaultHollowingState, persistActiveModelModifiers, scene.activeModel]);
-
-  const handleClearAppliedHollowing = React.useCallback(() => {
-    const activeModel = scene.activeModel;
-    if (!activeModel) return;
-
-    const sourceEntry = hollowingSourceByModelIdRef.current.get(activeModel.id)
-      ?? (() => {
-        const restored = geometryFromSnapshot(activeModel.meshModifiers?.hollowing ?? {});
-        if (!restored) return null;
-        const entry = { geometry: restored };
-        hollowingSourceByModelIdRef.current.set(activeModel.id, entry);
-        return entry;
-      })();
-
-    if (sourceEntry) {
-      const restoredGeometry = sourceEntry.geometry.clone();
-      const restored = scene.replaceModelGeometry(activeModel.id, restoredGeometry, 'Clear Hollowing');
-      if (!restored) {
-        restoredGeometry.dispose();
-      }
-    }
-
-    // Clear cavity geometry and disable interior view
-    const existingCavity = cavityGeometryByModelIdRef.current.get(activeModel.id);
-    if (existingCavity) {
-      existingCavity.geometry.dispose();
-      cavityGeometryByModelIdRef.current.delete(activeModel.id);
-    }
-    setInteriorView(false);
-
-    setHollowingDraftEnabled(false);
-    setHollowingEditMode(false);
-    setBlockedHollowVoxelIndices([]);
-    setEditingBlockedHollowVoxelIndices([]);
-    persistActiveModelModifiers({
-      ...(activeModel.meshModifiers ?? {}),
-      hollowing: {
-        enabled: false,
-        bakedIntoGeometry: false,
-        sourcePositionsBase64: undefined,
-        sourcePositionCount: undefined,
-        blockedVoxelIndices: [],
-        // Keep current settings — don't reset to defaults.
-        mode: hollowingState.mode,
-        voxelSizeMm: hollowingState.voxelSizeMm,
-        shellThicknessMm: hollowingState.shellThicknessMm,
-        infillMode: hollowingState.infillMode,
-        infillCellMm: hollowingState.infillCellMm,
-        infillBeamRadiusMm: hollowingState.infillBeamRadiusMm,
-        openFace: hollowingState.openFace,
-        openFaceSelected: hollowingState.mode === 'shell_open_face'
-          ? isShellOpenFaceSelected
-          : true,
-      },
-      holePunchAppliedPlacements: activeModel.meshModifiers?.holePunches ?? [],
-      holePunchesBakedIntoGeometry: activeModel.meshModifiers?.holePunchesBakedIntoGeometry === true,
-      holePunchSourcePositionsBase64: activeModel.meshModifiers?.holePunchSourcePositionsBase64,
-      holePunchSourcePositionCount: activeModel.meshModifiers?.holePunchSourcePositionCount,
-    });
-  }, [hollowingState, isShellOpenFaceSelected, persistActiveModelModifiers, scene.activeModel]);
-
-  const handleResetHollowingSettings = React.useCallback(() => {
-    setHollowingState(defaultHollowingState);
-    setIsShellOpenFaceSelected(true);
-  }, [defaultHollowingState]);
-
-  const handleHollowingStateChange = React.useCallback((next: HollowingPanelState) => {
-    const openFaceChanged = next.openFace !== hollowingState.openFace;
-    const resolutionChanged = Math.abs(next.voxelSizeMm - hollowingState.voxelSizeMm) > 1e-6;
-    const thicknessChanged = Math.abs(next.shellThicknessMm - hollowingState.shellThicknessMm) > 1e-6;
-    const blockedVoxelIndices = resolutionChanged
-      ? []
-      : blockedHollowVoxelIndices;
-    const nextShellOpenFaceSelected = next.mode === 'shell_open_face'
-      ? (
-        hollowingState.mode !== 'shell_open_face'
-          ? false
-          : (openFaceChanged ? true : isShellOpenFaceSelected)
-      )
-      : true;
-
-    // Warn before clearing blockers when adjusting resolution or thickness.
-    if ((resolutionChanged || thicknessChanged) && blockedHollowVoxelIndices.length > 0) {
-      setPendingBlockerResetState(next);
-      return;
-    }
-
-    setHollowingState(next);
-    setIsShellOpenFaceSelected(nextShellOpenFaceSelected);
-    setHollowingDraftEnabled(true);
-    setBlockedHollowVoxelIndices(blockedVoxelIndices);
-    // Clear editing indices when voxel resolution changes (new grid), but
-    // preserve them when only shell thickness or mode changes so the user
-    // stays in sphere edit mode with their current selection intact.
-    if (!hollowingEditMode || resolutionChanged) {
-      setEditingBlockedHollowVoxelIndices(blockedVoxelIndices);
-    }
-
-    if (!nextShellOpenFaceSelected) {
-      setSelectedHolePunchPlacementIds([]);
-      setHoveredHolePunchPlacementId(null);
-      setHolePunchHoverPlacement(null);
-    }
-
-    const activeModel = scene.activeModel;
-    if (!activeModel) return;
-
-    persistActiveModelModifiers({
-      ...(activeModel.meshModifiers ?? {}),
-      hollowing: {
-        enabled: true,
-        bakedIntoGeometry: false,
-        sourcePositionsBase64: activeModel.meshModifiers?.hollowing?.sourcePositionsBase64,
-        sourcePositionCount: activeModel.meshModifiers?.hollowing?.sourcePositionCount,
-        blockedVoxelIndices,
-        mode: next.mode,
-        voxelSizeMm: next.voxelSizeMm,
-        shellThicknessMm: next.shellThicknessMm,
-        infillMode: next.infillMode,
-        infillCellMm: next.infillCellMm,
-        infillBeamRadiusMm: next.infillBeamRadiusMm,
-        openFace: next.openFace,
-        openFaceSelected: nextShellOpenFaceSelected,
-      },
-    });
-  }, [blockedHollowVoxelIndices, hollowingState.mode, hollowingState.openFace, hollowingState.voxelSizeMm, isShellOpenFaceSelected, persistActiveModelModifiers, scene.activeModel]);
-
-
-  const isHollowingApplied = React.useMemo(() => {
-    const modifier = scene.activeModel?.meshModifiers?.hollowing;
-    return Boolean(modifier?.enabled && modifier?.bakedIntoGeometry);
-  }, [scene.activeModel]);
-
-
-
-
-
-
-
-
-
-
-
-
-
-  const persistedHollowingSignature = React.useMemo(
-    () => serializeHollowingModifier(scene.activeModel?.meshModifiers?.hollowing),
-    [scene.activeModel],
-  );
-
-  const draftHollowingSignature = React.useMemo(
-    () => serializeHollowingModifier({
-      enabled: hollowingDraftEnabled,
-      blockedVoxelIndices: blockedHollowVoxelIndices,
-      mode: hollowingState.mode,
-      voxelSizeMm: hollowingState.voxelSizeMm,
-      shellThicknessMm: hollowingState.shellThicknessMm,
-      infillMode: hollowingState.infillMode,
-      infillCellMm: hollowingState.infillCellMm,
-      infillBeamRadiusMm: hollowingState.infillBeamRadiusMm,
-      openFace: hollowingState.openFace,
-      openFaceSelected: hollowingState.mode === 'shell_open_face'
-        ? isShellOpenFaceSelected
-        : true,
-    }),
-    [
-      hollowingDraftEnabled,
-      blockedHollowVoxelIndices,
-      hollowingState.mode,
-      hollowingState.openFace,
-      hollowingState.infillMode,
-      hollowingState.shellThicknessMm,
-      hollowingState.infillBeamRadiusMm,
-      hollowingState.infillCellMm,
-      hollowingState.voxelSizeMm,
-      isShellOpenFaceSelected,
-    ],
-  );
-
-  const isShellFaceSelectionPending = hollowingState.mode === 'shell_open_face' && !isShellOpenFaceSelected;
-
-  const isHollowingDirty = draftHollowingSignature !== persistedHollowingSignature;
-
-  const canResetHollowing = React.useMemo(() => {
-    const activeModel = scene.activeModel;
-    if (!activeModel) return false;
-    const modifier = activeModel.meshModifiers?.hollowing;
-    if (!modifier) return false;
-    return Boolean(modifier.enabled || isHollowingDirty || isHollowingApplied);
-  }, [isHollowingApplied, isHollowingDirty, scene.activeModel]);
-
-  const blockedHollowVoxelIndexSet = React.useMemo(
-    () => new Set(blockedHollowVoxelIndices),
-    [blockedHollowVoxelIndices],
-  );
-
-  const editingBlockedHollowVoxelIndexSet = React.useMemo(
-    () => new Set(editingBlockedHollowVoxelIndices),
-    [editingBlockedHollowVoxelIndices],
-  );
-
-  React.useEffect(() => {
-    editingBlockedHollowVoxelIndicesRef.current = editingBlockedHollowVoxelIndices;
-  }, [editingBlockedHollowVoxelIndices]);
-
-  React.useEffect(() => {
-    hollowingEditModeRef.current = hollowingEditMode;
-  }, [hollowingEditMode]);
-
-  React.useEffect(() => {
-    if (hollowingEditMode) return;
-    hollowVoxelEditUndoStackRef.current = [];
-    hollowVoxelEditRedoStackRef.current = [];
-  }, [hollowingEditMode]);
-
-  const applyEditingBlockedHollowVoxelIndices = React.useCallback((
-    nextIndicesInput: Iterable<number>,
-    options?: { recordHistory?: boolean },
-  ) => {
-    const nextIndices = [...new Set(nextIndicesInput)]
-      .filter((value) => Number.isFinite(value))
-      .sort((a, b) => a - b);
-    const previousIndices = editingBlockedHollowVoxelIndicesRef.current;
-    if (areSortedNumberArraysEqual(previousIndices, nextIndices)) {
-      return false;
-    }
-
-    if (options?.recordHistory ?? true) {
-      hollowVoxelEditUndoStackRef.current.push([...previousIndices]);
-      if (hollowVoxelEditUndoStackRef.current.length > 100) {
-        hollowVoxelEditUndoStackRef.current.shift();
-      }
-      hollowVoxelEditRedoStackRef.current = [];
-    }
-
-    editingBlockedHollowVoxelIndicesRef.current = nextIndices;
-    setEditingBlockedHollowVoxelIndices(nextIndices);
-    return true;
-  }, []);
-
-  const undoHollowVoxelEdit = React.useCallback(() => {
-    const previousIndices = hollowVoxelEditUndoStackRef.current.pop();
-    if (!previousIndices) return false;
-
-    hollowVoxelEditRedoStackRef.current.push([...editingBlockedHollowVoxelIndicesRef.current]);
-    editingBlockedHollowVoxelIndicesRef.current = previousIndices;
-    setEditingBlockedHollowVoxelIndices(previousIndices);
-    return true;
-  }, []);
-
-  const redoHollowVoxelEdit = React.useCallback(() => {
-    const nextIndices = hollowVoxelEditRedoStackRef.current.pop();
-    if (!nextIndices) return false;
-
-    hollowVoxelEditUndoStackRef.current.push([...editingBlockedHollowVoxelIndicesRef.current]);
-    editingBlockedHollowVoxelIndicesRef.current = nextIndices;
-    setEditingBlockedHollowVoxelIndices(nextIndices);
-    return true;
-  }, []);
-
-  React.useEffect(() => {
-    if (scene.mode !== 'prepare' || transformMgr.transformMode !== 'hollowing' || !hollowingEditMode) {
-      return;
-    }
-
-    const handleHollowVoxelEditHistoryHotkey = (event: KeyboardEvent) => {
-      if (isKeyboardTargetEditable(event.target)) return;
-      if (!(event.ctrlKey || event.metaKey)) return;
-
-      const key = event.key.toLowerCase();
-      const handled = key === 'y'
-        ? redoHollowVoxelEdit()
-        : key === 'z'
-          ? (event.shiftKey ? redoHollowVoxelEdit() : undoHollowVoxelEdit())
-          : false;
-
-      if (!handled) return;
-
-      event.preventDefault();
-      event.stopPropagation();
-      event.stopImmediatePropagation?.();
-    };
-
-    window.addEventListener('keydown', handleHollowVoxelEditHistoryHotkey, true);
-    return () => {
-      window.removeEventListener('keydown', handleHollowVoxelEditHistoryHotkey, true);
-    };
-  }, [hollowingEditMode, redoHollowVoxelEdit, scene.mode, transformMgr.transformMode, undoHollowVoxelEdit]);
-
-  const blockedPreviewVoxelInstanceIdSet = React.useMemo(() => {
-    const preview = hollowPreview;
-    if (!preview) return new Set<number>();
-    const activeBlockedIndexSet = hollowingEditMode
-      ? editingBlockedHollowVoxelIndexSet
-      : blockedHollowVoxelIndexSet;
-
-    const next = new Set<number>();
-    for (let instanceIndex = 0; instanceIndex < preview.removedVoxelIndices.length; instanceIndex += 1) {
-      if (activeBlockedIndexSet.has(preview.removedVoxelIndices[instanceIndex] ?? -1)) {
-        next.add(instanceIndex);
-      }
-    }
-    // Also map committed blocked voxel centers: their instance index is
-    // offset past the removed voxels. If the user cleared a blocked voxel
-    // from the editing set, it won't be in activeBlockedIndexSet and will
-    // render as yellow.
-    if (preview.blockedVoxelCenters) {
-      const blockedCount = Math.floor(preview.blockedVoxelCenters.length / 3);
-      for (let blockedIndex = 0; blockedIndex < blockedCount; blockedIndex += 1) {
-        const gridIndex = blockedHollowVoxelIndices[blockedIndex];
-        if (activeBlockedIndexSet.has(gridIndex)) {
-          next.add(preview.removedVoxelIndices.length + blockedIndex);
-        }
-      }
-    }
-    return next;
-  }, [blockedHollowVoxelIndexSet, blockedHollowVoxelIndices, editingBlockedHollowVoxelIndexSet, hollowPreview, hollowingEditMode]);
-
-  const commitBlockedHollowVoxelIndices = React.useCallback((nextIndices: number[]) => {
-    const activeModel = scene.activeModel;
-    if (!activeModel) return;
-
-    setBlockedHollowVoxelIndices(nextIndices);
-    setHollowingDraftEnabled(true);
-    persistActiveModelModifiers({
-      ...(activeModel.meshModifiers ?? {}),
-      hollowing: {
-        enabled: true,
-        bakedIntoGeometry: false,
-        sourcePositionsBase64: activeModel.meshModifiers?.hollowing?.sourcePositionsBase64,
-        sourcePositionCount: activeModel.meshModifiers?.hollowing?.sourcePositionCount,
-        blockedVoxelIndices: nextIndices,
-        mode: hollowingState.mode,
-        voxelSizeMm: hollowingState.voxelSizeMm,
-        shellThicknessMm: hollowingState.shellThicknessMm,
-        infillMode: hollowingState.infillMode,
-        infillCellMm: hollowingState.infillCellMm,
-        infillBeamRadiusMm: hollowingState.infillBeamRadiusMm,
-        openFace: hollowingState.openFace,
-        openFaceSelected: hollowingState.mode === 'shell_open_face'
-          ? isShellOpenFaceSelected
-          : true,
-      },
-    });
-  }, [hollowingState, isShellOpenFaceSelected, persistActiveModelModifiers, scene.activeModel]);
-
-  const toggleBlockedHollowVoxelIndex = React.useCallback((voxelIndex: number) => {
-    const currentPreview = hollowPreview;
-    if (!currentPreview || voxelIndex < 0) return;
-
-    let gridVoxelIndex: number;
-
-    const removedCount = currentPreview.removedVoxelIndices.length;
-    if (voxelIndex < removedCount) {
-      // Instance in the removed voxel array — look up grid index directly.
-      gridVoxelIndex = currentPreview.removedVoxelIndices[voxelIndex];
-    } else {
-      // Instance in the appended blocked-only array — look up via
-      // blockedHollowVoxelIndices (committed set, same order as blocked centers).
-      const blockedOffset = voxelIndex - removedCount;
-      gridVoxelIndex = blockedHollowVoxelIndices[blockedOffset];
-    }
-
-    if (!Number.isFinite(gridVoxelIndex)) return;
-
-    const next = new Set(editingBlockedHollowVoxelIndexSet);
-    if (next.has(gridVoxelIndex)) {
-      next.delete(gridVoxelIndex);
-    } else {
-      next.add(gridVoxelIndex);
-    }
-    applyEditingBlockedHollowVoxelIndices(next);
-  }, [applyEditingBlockedHollowVoxelIndices, blockedHollowVoxelIndices, editingBlockedHollowVoxelIndexSet, hollowPreview]);
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-  const requestResetHollowing = React.useCallback(() => {
-    if (!canResetHollowing || isApplyingHollowing || isPreviewingHollowing) return;
-    setPendingModifierResetAction('hollowing');
-  }, [canResetHollowing, isApplyingHollowing, isPreviewingHollowing]);
-
-  const requestClearAppliedHollowing = React.useCallback(() => {
-    setPendingModifierResetAction('clear_hollowing');
-  }, []);
-
+  // Relocated from the early state block: depends on hollowPreview which is now
+  // produced by useHollowingManager (declared above, after transformMgr).
+  const shouldForceHollowingXray = scene.mode === 'prepare'
+    && transformMgr.transformMode === 'hollowing'
+    && !scene.activeModel?.meshModifiers?.hollowing?.bakedIntoGeometry;
+  const effectiveShaderType = (shouldForceHollowingXray || hollowPreview)
+    ? 'xray'
+    : (sessionShaderOverride ?? scene.shaderType);
+
+  // Populate the hollowing manager deps now that the hole-punch manager and
+  // shared callbacks exist (breaks the TDZ/dependency cycle).
+  hollowingDepsRef.current = {
+    showOperationError,
+    setShowDamagedModelDialog,
+    persistActiveModelModifiers,
+    setPendingModifierResetAction,
+    setInteriorView,
+    setSessionShaderOverride,
+    computeAutoHolePunchDepthMmForGeometry,
+    setHolePunchState,
+    setHolePunchPlacements,
+    holePunchPlacementsRef,
+    setPendingHolePunchAutoApplyModelId,
+    setPendingBlockerResetState,
+    setSelectedHolePunchPlacementIds,
+    setHoveredHolePunchPlacementId,
+    setHolePunchHoverPlacement,
+    interiorView,
+  };
 
   const handleConfirmModifierReset = React.useCallback(() => {
     const action = pendingModifierResetAction;
@@ -13751,248 +13139,6 @@ export default function Home() {
 
 
 
-  const clearPendingHollowPreviewDebounce = React.useCallback(() => {
-    if (hollowPreviewDebounceTimerRef.current !== null) {
-      clearTimeout(hollowPreviewDebounceTimerRef.current);
-      hollowPreviewDebounceTimerRef.current = null;
-    }
-  }, []);
-
-  const resolveHollowPreviewSourceGeometry = React.useCallback((activeModel: (typeof scene.models)[number]) => {
-    const sourceEntry = hollowingSourceByModelIdRef.current.get(activeModel.id);
-    if (sourceEntry) {
-      return sourceEntry.geometry;
-    }
-
-    // Only restore from the hollowing snapshot if hollowing is actually baked
-    // (or at least enabled). If hollowing was reset/cleared, the snapshot is a
-    // stale copy of the pre-hollowing geometry which may have holes that have
-    // since been removed — using it would make the preview ignore hole changes.
-    const h = activeModel.meshModifiers?.hollowing;
-    const snapshotIsValid = h?.sourcePositionsBase64 && (h.bakedIntoGeometry || h.enabled);
-    const restoredFromSnapshot = snapshotIsValid
-      ? geometryFromSnapshot(h)
-      : null;
-    if (restoredFromSnapshot) {
-      hollowingSourceByModelIdRef.current.set(activeModel.id, { geometry: restoredFromSnapshot });
-      return restoredFromSnapshot;
-    }
-
-    return activeModel.geometry.geometry;
-  }, []);
-
-  const buildHollowingOptions = React.useCallback((
-    modelScale: THREE.Vector3,
-    maxExtent: number,
-    tuning?: { preview?: boolean; previewShellThicknessMm?: number },
-    stateOverride?: HollowingPanelState,
-  ): HollowOptions => {
-    const preview = Boolean(tuning?.preview);
-    const state = stateOverride ?? hollowingState;
-    const effectiveHollowMode = state.mode === 'shell_open_face'
-      ? 'cavity'
-      : state.mode;
-    const voxelResolution = computeVoxelResolution(
-      worldMmToLocalMm(state.voxelSizeMm, getUniformScaleFactorForThickness(modelScale)),
-      maxExtent,
-    );
-    const shellThicknessMmWorld = preview
-      ? (tuning?.previewShellThicknessMm ?? state.shellThicknessMm)
-      : state.shellThicknessMm;
-    const hasCommittedBlockedVoxels = blockedHollowVoxelIndices.length > 0;
-
-    return {
-      mode: effectiveHollowMode,
-      voxelResolution,
-      shellThicknessMm: worldMmToLocalMm(
-        shellThicknessMmWorld,
-        getUniformScaleFactorForThickness(modelScale),
-      ),
-      blockedVoxelIndices: blockedHollowVoxelIndices,
-      infillMode: state.infillMode,
-      infillCellMm: worldMmToLocalMm(
-        state.infillCellMm,
-        getUniformScaleFactorForThickness(modelScale),
-      ),
-      infillBeamRadiusMm: worldMmToLocalMm(
-        state.infillBeamRadiusMm,
-        getUniformScaleFactorForThickness(modelScale),
-      ),
-      openFace: state.openFace,
-      drainHoles: [],
-      previewCavityOnly: false,
-      smoothInternalSurfaces: !preview || hasCommittedBlockedVoxels,
-      internalChamferPasses: !preview || hasCommittedBlockedVoxels ? 2 : 0,
-    };
-  }, [
-    hollowingState.mode,
-    hollowingState.openFace,
-    hollowingState.infillMode,
-    hollowingState.infillBeamRadiusMm,
-    hollowingState.infillCellMm,
-    hollowingState.shellThicknessMm,
-    hollowingState.voxelSizeMm,
-    blockedHollowVoxelIndices,
-  ]);
-
-  const buildHollowPreviewRequest = React.useCallback((
-    activeModel: (typeof scene.models)[number],
-    overrideState?: HollowingPanelState,
-  ) => {
-    const previewState = overrideState ?? hollowingState;
-    const previewShellThicknessMm = quantizePreviewShellThicknessMm(previewState.shellThicknessMm);
-    const sourceGeometry = resolveHollowPreviewSourceGeometry(activeModel);
-    const sourceGeometryKey = buildGeometryVersionKey(sourceGeometry);
-    const bbox = sourceGeometry.boundingBox ?? new THREE.Box3().setFromBufferAttribute(
-      sourceGeometry.getAttribute('position') as THREE.BufferAttribute,
-    );
-    const bboxSize = bbox.getSize(new THREE.Vector3());
-    const maxExtent = Math.max(bboxSize.x, bboxSize.y, bboxSize.z);
-    const previewQuat = new THREE.Quaternion().setFromEuler(activeModel.transform.rotation);
-    const options: HollowOptions = {
-      ...buildHollowingOptions(activeModel.transform.scale, maxExtent, {
-        preview: true,
-        previewShellThicknessMm,
-      }, previewState),
-      drainHoles: [],
-      previewCavityOnly: true,
-      previewVoxelSpheres: true,
-      rotationQuat: [previewQuat.x, previewQuat.y, previewQuat.z, previewQuat.w],
-    };
-    const optionsKey = JSON.stringify(options);
-    const previewKey = `${activeModel.id}::${sourceGeometryKey}::${optionsKey}`;
-
-    return {
-      sourceGeometry,
-      sourceGeometryKey,
-      options,
-      previewKey,
-    };
-  }, [
-    buildHollowingOptions,
-    hollowingState.shellThicknessMm,
-    hollowingState.voxelSizeMm,
-    resolveHollowPreviewSourceGeometry,
-  ]);
-
-  const cacheHollowPreviewResult = React.useCallback((
-    activeModelId: string,
-    report: HollowReport,
-    positions: Float32Array,
-    infillPositions: Float32Array | undefined,
-    removedVoxelCenters: Float32Array | undefined,
-    removedVoxelIndices: Uint32Array | undefined,
-    blockedVoxelCenters: Float32Array | undefined,
-    previewKey: string,
-  ) => {
-    const cachedPositions = new Float32Array(positions.length);
-    cachedPositions.set(positions);
-    const cachedRemovedVoxelCenters = removedVoxelCenters
-      ? new Float32Array(removedVoxelCenters)
-      : undefined;
-    const cachedRemovedVoxelIndices = removedVoxelIndices
-      ? new Uint32Array(removedVoxelIndices)
-      : undefined;
-    const cachedBlockedVoxelCenters = blockedVoxelCenters
-      ? new Float32Array(blockedVoxelCenters)
-      : undefined;
-
-    hollowPreviewResultCacheRef.current.set(previewKey, {
-      modelId: activeModelId,
-      report,
-      positions: cachedPositions,
-      infillPositions,
-      removedVoxelCenters: cachedRemovedVoxelCenters,
-      removedVoxelIndices: cachedRemovedVoxelIndices,
-      blockedVoxelCenters: cachedBlockedVoxelCenters,
-      previewGeometry: null,
-      infillGeometry: null,
-    });
-
-    if (hollowPreviewResultCacheRef.current.size > 6) {
-      const oldest = hollowPreviewResultCacheRef.current.keys().next().value;
-      if (oldest != null) {
-        const evicted = hollowPreviewResultCacheRef.current.get(oldest);
-        if (evicted) {
-          disposeHollowPreviewCacheEntry(evicted);
-        }
-        hollowPreviewResultCacheRef.current.delete(oldest);
-      }
-    }
-
-    return cachedPositions;
-  }, []);
-
-  const materializeHollowPreviewCacheEntry = React.useCallback((previewKey: string) => {
-    const cached = hollowPreviewResultCacheRef.current.get(previewKey);
-    if (!cached) return null;
-
-    if (!cached.previewGeometry) {
-      cached.previewGeometry = createGeometryFromPreviewPositions(cached.positions);
-    }
-
-    if (cached.infillPositions && !cached.infillGeometry) {
-      cached.infillGeometry = createGeometryFromPreviewPositions(cached.infillPositions);
-    }
-
-    return cached;
-  }, []);
-
-  const primeHollowPreviewCache = React.useCallback(async (
-    activeModel: (typeof scene.models)[number],
-    overrideState?: HollowingPanelState,
-  ) => {
-    const { sourceGeometry, sourceGeometryKey, options, previewKey } = buildHollowPreviewRequest(activeModel, overrideState);
-
-    if (hollowPreviewResultCacheRef.current.has(previewKey) || hollowPreviewWarmupKeyRef.current === previewKey) {
-      return;
-    }
-
-    hollowPreviewWarmupKeyRef.current = previewKey;
-    try {
-      const staged = await stageHollowPreviewSource(
-        sourceGeometry,
-        `${activeModel.id}::${sourceGeometryKey}`,
-      );
-      if (!staged) {
-        return;
-      }
-
-      const result = await hollowPreviewFromCapturedSource(options);
-      if (!result) {
-        return;
-      }
-
-      cacheHollowPreviewResult(
-        activeModel.id,
-        result.report,
-        result.positions,
-        result.infillPositions,
-        result.removedVoxelCenters,
-        result.removedVoxelIndices,
-        result.blockedVoxelCenters,
-        previewKey,
-      );
-
-      const scheduleMaterialize = typeof window !== 'undefined' && typeof window.requestIdleCallback === 'function'
-        ? (cb: () => void) => window.requestIdleCallback(() => cb())
-        : (cb: () => void) => window.setTimeout(cb, 0);
-      scheduleMaterialize(() => {
-        try {
-          materializeHollowPreviewCacheEntry(previewKey);
-        } catch (error) {
-          console.warn('[Hollowing] Failed to materialize cached preview geometry:', error);
-        }
-      });
-    } catch (error) {
-      console.warn('[Hollowing] Warm preview prime failed:', error);
-    } finally {
-      if (hollowPreviewWarmupKeyRef.current === previewKey) {
-        hollowPreviewWarmupKeyRef.current = null;
-      }
-    }
-  }, [buildHollowPreviewRequest, cacheHollowPreviewResult, materializeHollowPreviewCacheEntry]);
-
   const handleTransformToolbarHover = React.useCallback((mode: TransformMode | null) => {
     if (mode === 'hollowing') {
       if (scene.mode === 'prepare') {
@@ -14023,377 +13169,6 @@ export default function Home() {
     scene.mode,
   ]);
 
-  const runHollowPreview = React.useCallback(async ({
-    activeModel,
-    sourceGeometry,
-    sourceGeometryKey,
-    options,
-    previewKey,
-    notifyUnavailable,
-  }: {
-    activeModel: (typeof scene.models)[number];
-    sourceGeometry: THREE.BufferGeometry;
-    sourceGeometryKey: string;
-    options: HollowOptions;
-    previewKey: string;
-    notifyUnavailable: boolean;
-  }) => {
-    const requestSeq = ++hollowPreviewRequestSeqRef.current;
-    setIsPreviewingHollowing(true);
-
-    try {
-      const cached = hollowPreviewResultCacheRef.current.get(previewKey);
-      if (cached) {
-        hollowPreviewResultCacheRef.current.delete(previewKey);
-        hollowPreviewResultCacheRef.current.set(previewKey, cached);
-        const materialized = materializeHollowPreviewCacheEntry(previewKey) ?? cached;
-        const previewGeometry = materialized.previewGeometry ?? createGeometryFromPreviewPositions(materialized.positions);
-        const infillGeometry = materialized.infillGeometry
-          ?? (materialized.infillPositions ? createGeometryFromPreviewPositions(materialized.infillPositions) : null);
-        if (hollowPreviewRequestSeqRef.current !== requestSeq) {
-          disposeHollowPreviewGeometryIfUncached(previewGeometry, hollowPreviewResultCacheRef.current.values());
-          disposeHollowPreviewGeometryIfUncached(infillGeometry, hollowPreviewResultCacheRef.current.values());
-          return;
-        }
-
-        setHollowPreview((previous) => {
-          if (previous) {
-            disposeHollowPreviewGeometryIfUncached(previous.geometry, hollowPreviewResultCacheRef.current.values());
-            disposeHollowPreviewGeometryIfUncached(previous.infillGeometry ?? null, hollowPreviewResultCacheRef.current.values());
-          }
-          return {
-            modelId: cached.modelId,
-            geometry: previewGeometry,
-            infillGeometry,
-            removedVoxelCenters: cached.removedVoxelCenters ?? new Float32Array(0),
-            removedVoxelIndices: cached.removedVoxelIndices ?? new Uint32Array(0),
-            blockedVoxelCenters: cached.blockedVoxelCenters,
-            report: cached.report,
-            previewKey,
-            previewVoxelSpheres: true,
-          };
-        });
-        return;
-      }
-
-      const staged = await stageHollowPreviewSource(
-        sourceGeometry,
-        `${activeModel.id}::${sourceGeometryKey}`,
-      );
-      if (!staged) {
-        if (notifyUnavailable) {
-          showOperationError('Hollowing preview is available in DragonFruit Desktop only.');
-        }
-        return;
-      }
-
-      const result = await hollowPreviewFromCapturedSource(options);
-      if (!result) {
-        if (notifyUnavailable) {
-          showOperationError('Hollowing preview is available in DragonFruit Desktop only.');
-        }
-        return;
-      }
-
-      const cachedPositions = cacheHollowPreviewResult(
-        activeModel.id,
-        result.report,
-        result.positions,
-        result.infillPositions,
-        result.removedVoxelCenters,
-        result.removedVoxelIndices,
-        result.blockedVoxelCenters,
-        previewKey,
-      );
-      const materialized = materializeHollowPreviewCacheEntry(previewKey);
-
-      const previewGeometry = materialized?.previewGeometry
-        ?? createGeometryFromPreviewPositions(cachedPositions);
-      const infillGeometry = materialized?.infillGeometry
-        ?? (result.infillPositions
-          ? createGeometryFromPreviewPositions(result.infillPositions)
-          : null);
-
-      if (hollowPreviewRequestSeqRef.current !== requestSeq) {
-        disposeHollowPreviewGeometryIfUncached(previewGeometry, hollowPreviewResultCacheRef.current.values());
-        disposeHollowPreviewGeometryIfUncached(infillGeometry, hollowPreviewResultCacheRef.current.values());
-        return;
-      }
-
-      setHollowPreview((previous) => {
-        if (previous) {
-          disposeHollowPreviewGeometryIfUncached(previous.geometry, hollowPreviewResultCacheRef.current.values());
-          disposeHollowPreviewGeometryIfUncached(previous.infillGeometry ?? null, hollowPreviewResultCacheRef.current.values());
-        }
-        return {
-          modelId: activeModel.id,
-          geometry: previewGeometry,
-          infillGeometry,
-          removedVoxelCenters: result.removedVoxelCenters ?? new Float32Array(0),
-          removedVoxelIndices: result.removedVoxelIndices ?? new Uint32Array(0),
-          blockedVoxelCenters: result.blockedVoxelCenters,
-          report: result.report,
-          previewKey,
-          previewVoxelSpheres: true,
-        };
-      });
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      if (notifyUnavailable) {
-        showOperationError(`Hollowing preview failed: ${message}`);
-      } else {
-        console.warn('[Hollowing] Debounced preview failed:', message);
-      }
-    } finally {
-      if (hollowPreviewRequestSeqRef.current === requestSeq) {
-        setIsPreviewingHollowing(false);
-        setIsApplyingBlockersHollowing(false);
-      }
-    }
-  }, [cacheHollowPreviewResult, scene.models]);
-
-  const clearHollowPreview = React.useCallback(() => {
-    hollowPreviewRequestSeqRef.current += 1;
-    setIsPreviewingHollowing(false);
-    clearPendingHollowPreviewDebounce();
-    setHollowPreview((previous) => {
-      if (previous) {
-        disposeHollowPreviewGeometryIfUncached(previous.geometry, hollowPreviewResultCacheRef.current.values());
-        disposeHollowPreviewGeometryIfUncached(previous.infillGeometry ?? null, hollowPreviewResultCacheRef.current.values());
-      }
-      return null;
-    });
-  }, [clearPendingHollowPreviewDebounce]);
-
-  React.useEffect(() => {
-    return () => {
-      if (hollowPreview) {
-        hollowPreview.geometry.dispose();
-        hollowPreview.infillGeometry?.dispose();
-      }
-    };
-  }, [hollowPreview]);
-
-  React.useEffect(() => {
-    return () => {
-      clearPendingHollowPreviewDebounce();
-    };
-  }, [clearPendingHollowPreviewDebounce]);
-
-  React.useEffect(() => {
-    const liveIds = new Set(scene.models.map((model) => model.id));
-    for (const [modelId, entry] of hollowingSourceByModelIdRef.current.entries()) {
-      if (liveIds.has(modelId)) continue;
-      entry.geometry.dispose();
-      hollowingSourceByModelIdRef.current.delete(modelId);
-    }
-
-    for (const [modelId, entry] of cavityGeometryByModelIdRef.current.entries()) {
-      if (liveIds.has(modelId)) continue;
-      entry.geometry.dispose();
-      cavityGeometryByModelIdRef.current.delete(modelId);
-    }
-
-    // If the active model's cavity geometry was just removed, exit interior view
-    // so the user doesn't get stuck with no way to toggle it off.
-    if (
-      interiorView &&
-      (!scene.activeModel || !cavityGeometryByModelIdRef.current.has(scene.activeModel.id))
-    ) {
-      setInteriorView(false);
-    }
-
-    for (const [cacheKey, entry] of hollowPreviewResultCacheRef.current.entries()) {
-      if (liveIds.has(entry.modelId)) continue;
-      disposeHollowPreviewCacheEntry(entry);
-      hollowPreviewResultCacheRef.current.delete(cacheKey);
-    }
-  }, [scene.models, interiorView, setInteriorView]);
-
-  // Restore cavity geometry from persisted data for models with baked hollowing.
-  React.useEffect(() => {
-    for (const model of scene.models) {
-      const hollowing = model.meshModifiers?.hollowing;
-      if (!hollowing?.enabled || !hollowing.cavityPositionsBase64 || !hollowing.cavityPositionCount) {
-        continue;
-      }
-      if (cavityGeometryByModelIdRef.current.has(model.id)) {
-        continue; // already restored
-      }
-
-      const bytes = base64ToBytes(hollowing.cavityPositionsBase64);
-      if (bytes.byteLength % Float32Array.BYTES_PER_ELEMENT !== 0) continue;
-      const view = new Float32Array(
-        bytes.buffer,
-        bytes.byteOffset,
-        bytes.byteLength / Float32Array.BYTES_PER_ELEMENT,
-      );
-      if (view.length !== hollowing.cavityPositionCount * 3) continue;
-
-      const positions = new Float32Array(view.length);
-      positions.set(view);
-      const cavityGeometry = new THREE.BufferGeometry();
-      cavityGeometry.setAttribute('position', new THREE.BufferAttribute(positions, 3));
-      cavityGeometry.computeVertexNormals();
-      cavityGeometry.computeBoundingBox();
-      cavityGeometry.computeBoundingSphere();
-      cavityGeometryByModelIdRef.current.set(model.id, { geometry: cavityGeometry });
-    }
-  }, [scene.models]);
-
-  React.useEffect(() => {
-    return () => {
-      for (const entry of hollowingSourceByModelIdRef.current.values()) {
-        entry.geometry.dispose();
-      }
-      hollowingSourceByModelIdRef.current.clear();
-      for (const entry of cavityGeometryByModelIdRef.current.values()) {
-        entry.geometry.dispose();
-      }
-      cavityGeometryByModelIdRef.current.clear();
-      for (const entry of hollowPreviewResultCacheRef.current.values()) {
-        disposeHollowPreviewCacheEntry(entry);
-      }
-      hollowPreviewResultCacheRef.current.clear();
-    };
-  }, []);
-
-  React.useEffect(() => {
-    if (!hollowPreview) return;
-    if (scene.mode !== 'prepare' || transformMgr.transformMode !== 'hollowing') {
-      clearHollowPreview();
-      return;
-    }
-    const stillExists = scene.models.some((model) => model.id === hollowPreview.modelId);
-    if (!stillExists) {
-      clearHollowPreview();
-    }
-  }, [clearHollowPreview, hollowPreview, scene.mode, scene.models, transformMgr.transformMode]);
-
-  React.useEffect(() => {
-    if (scene.mode === 'prepare' && transformMgr.transformMode === 'hollowing') {
-      return;
-    }
-    setHollowingEditMode(false);
-    setEditingBlockedHollowVoxelIndices(blockedHollowVoxelIndices);
-  }, [blockedHollowVoxelIndices, scene.mode, transformMgr.transformMode]);
-
-  const resolveBlockedHollowVoxelMarqueeSelection = React.useCallback((
-    polygon: Array<{ x: number; y: number }>,
-    helpers: {
-      projectWorldPoint: (point: THREE.Vector3) => { x: number; y: number; z: number } | null;
-    },
-  ) => {
-    const preview = hollowPreview;
-    const activeModel = scene.activeModel;
-    if (!preview || !activeModel || polygon.length < 3) return [] as string[];
-
-    const pointInPolygon = (x: number, y: number) => {
-      let inside = false;
-      for (let i = 0, j = polygon.length - 1; i < polygon.length; j = i, i += 1) {
-        const xi = polygon[i].x;
-        const yi = polygon[i].y;
-        const xj = polygon[j].x;
-        const yj = polygon[j].y;
-        const intersects = ((yi > y) !== (yj > y))
-          && (x < ((xj - xi) * (y - yi)) / ((yj - yi) || 1e-6) + xi);
-        if (intersects) inside = !inside;
-      }
-      return inside;
-    };
-
-    const modelQuaternion = quaternionFromGlobalEuler(activeModel.transform.rotation);
-    const selected: string[] = [];
-
-    for (let instanceIndex = 0; instanceIndex < preview.removedVoxelIndices.length; instanceIndex += 1) {
-      const offset = instanceIndex * 3;
-      const localPoint = new THREE.Vector3(
-        preview.removedVoxelCenters[offset] - activeModel.geometry.center.x,
-        preview.removedVoxelCenters[offset + 1] - activeModel.geometry.center.y,
-        preview.removedVoxelCenters[offset + 2] - activeModel.geometry.center.z,
-      );
-      localPoint.multiply(activeModel.transform.scale);
-      localPoint.applyQuaternion(modelQuaternion);
-      localPoint.add(activeModel.transform.position);
-      const projected = helpers.projectWorldPoint(localPoint);
-      if (!projected) continue;
-      if (!pointInPolygon(projected.x, projected.y)) continue;
-      selected.push(String(preview.removedVoxelIndices[instanceIndex]));
-    }
-
-    // Also test blocked-only voxel centers so lassoing over them works.
-    if (preview.blockedVoxelCenters) {
-      const blockedCount = Math.floor(preview.blockedVoxelCenters.length / 3);
-      for (let blockedIndex = 0; blockedIndex < blockedCount; blockedIndex += 1) {
-        const offset = blockedIndex * 3;
-        const localPoint = new THREE.Vector3(
-          preview.blockedVoxelCenters[offset] - activeModel.geometry.center.x,
-          preview.blockedVoxelCenters[offset + 1] - activeModel.geometry.center.y,
-          preview.blockedVoxelCenters[offset + 2] - activeModel.geometry.center.z,
-        );
-        localPoint.multiply(activeModel.transform.scale);
-        localPoint.applyQuaternion(modelQuaternion);
-        localPoint.add(activeModel.transform.position);
-        const projected = helpers.projectWorldPoint(localPoint);
-        if (!projected) continue;
-        if (!pointInPolygon(projected.x, projected.y)) continue;
-        selected.push(String(blockedHollowVoxelIndices[blockedIndex]));
-      }
-    }
-
-    return selected;
-  }, [hollowPreview, scene.activeModel, blockedHollowVoxelIndices]);
-
-  const handleBlockedHollowVoxelMarqueeSelection = React.useCallback((ids: string[], altKey?: boolean) => {
-    if (ids.length === 0) return;
-    const next = new Set(editingBlockedHollowVoxelIndicesRef.current);
-    if (altKey) {
-      // Alt + lasso: un-block (remove from the blocked set).
-      for (const id of ids) {
-        const voxelIndex = Number(id);
-        if (!Number.isFinite(voxelIndex)) continue;
-        next.delete(voxelIndex);
-      }
-    } else {
-      // Plain lasso: block (add to the blocked set).
-      for (const id of ids) {
-        const voxelIndex = Number(id);
-        if (!Number.isFinite(voxelIndex)) continue;
-        next.add(voxelIndex);
-      }
-    }
-    applyEditingBlockedHollowVoxelIndices(next);
-  }, [applyEditingBlockedHollowVoxelIndices]);
-
-  const handleStartHollowVoxelEditing = React.useCallback(() => {
-    editingBlockedHollowVoxelIndicesRef.current = blockedHollowVoxelIndices;
-    hollowVoxelEditUndoStackRef.current = [];
-    hollowVoxelEditRedoStackRef.current = [];
-    setEditingBlockedHollowVoxelIndices(blockedHollowVoxelIndices);
-    setHollowingEditMode(true);
-  }, [blockedHollowVoxelIndices]);
-
-  const handleClearHollowVoxelEditing = React.useCallback(() => {
-    applyEditingBlockedHollowVoxelIndices([]);
-  }, [applyEditingBlockedHollowVoxelIndices]);
-
-  const handleDoneHollowVoxelEditing = React.useCallback(() => {
-    const nextIndices = [...editingBlockedHollowVoxelIndices].sort((a, b) => a - b);
-    const prevIndices = [...blockedHollowVoxelIndices].sort((a, b) => a - b);
-    const hasChanges = nextIndices.length !== prevIndices.length
-      || nextIndices.some((v, i) => v !== prevIndices[i]);
-
-    if (!hasChanges) {
-      setHollowingEditMode(false);
-      return;
-    }
-
-    commitBlockedHollowVoxelIndices(nextIndices);
-    clearHollowPreview();
-    setHollowingEditMode(false);
-    setIsApplyingBlockersHollowing(true);
-  }, [blockedHollowVoxelIndices, clearHollowPreview, commitBlockedHollowVoxelIndices, editingBlockedHollowVoxelIndices]);
-
   React.useEffect(() => {
     if (canUseAutoHolePunchDepth || holePunchState.depthMode !== 'auto') {
       return;
@@ -14401,45 +13176,6 @@ export default function Home() {
 
     setHolePunchState((previous) => ({ ...previous, depthMode: 'manual' }));
   }, [canUseAutoHolePunchDepth, holePunchState.depthMode]);
-
-  React.useEffect(() => {
-    if (scene.mode !== 'prepare' || transformMgr.transformMode === 'hollowing') {
-      return;
-    }
-
-    const activeModel = scene.activeModel;
-    if (!activeModel) {
-      return;
-    }
-
-    const persistedHollowing = activeModel.meshModifiers?.hollowing;
-    const warmupState: HollowingPanelState = persistedHollowing?.enabled
-      ? {
-          mode: persistedHollowing.mode === 'shell_open_face' ? 'cavity' : persistedHollowing.mode,
-          voxelSizeMm: persistedHollowing.voxelSizeMm,
-          shellThicknessMm: persistedHollowing.shellThicknessMm,
-          infillMode: persistedHollowing.infillMode ?? defaultHollowingState.infillMode,
-          infillCellMm: persistedHollowing.infillCellMm ?? defaultHollowingState.infillCellMm,
-          infillBeamRadiusMm: persistedHollowing.infillBeamRadiusMm ?? defaultHollowingState.infillBeamRadiusMm,
-          openFace: persistedHollowing.openFace,
-        }
-      : defaultHollowingState;
-
-    const previewRequest = buildHollowPreviewRequest(activeModel, warmupState);
-    if (hollowPreviewResultCacheRef.current.has(previewRequest.previewKey)
-      || hollowPreviewWarmupKeyRef.current === previewRequest.previewKey) {
-      return;
-    }
-
-    void primeHollowPreviewCache(activeModel, warmupState);
-  }, [
-    buildHollowPreviewRequest,
-    defaultHollowingState,
-    primeHollowPreviewCache,
-    scene.activeModel,
-    scene.mode,
-    transformMgr.transformMode,
-  ]);
 
   React.useEffect(() => {
     const activeModel = scene.activeModel;
